+++ src/AStarSearch.cpp
@@ -49,7 +49,7 @@ void AStarSearch::Search(Map* pMap, UnitBase* pUnit, Coord start, Coord destinat
 
     const auto destinationTile = pMap->getTile(destination);
 
-    const FixPoint rotationSpeed = 1_fix / (currentGame->objectData.data[pUnit->getItemID()][pUnit->getOriginalHouseID()].turnspeed * TILESIZE);
+    const FixPoint rotationSpeed = 1_fix / (currentGame->objectData.data[pUnit->getItemID()][static_cast<int>(pUnit->getOriginalHouseID())].turnspeed * TILESIZE);
 
     const auto heuristic = blockDistance(start, destination);
     auto smallestHeuristic = FixPt_MAX;
