+++ src/players/SmartBot.cpp
@@ -88,7 +88,7 @@ void SmartBot::save(OutputStream& stream) const {
 
 
 void SmartBot::update() {
-    if( (getGameCycleCount() + getHouse()->getHouseID()) % AIUPDATEINTERVAL != 0) {
+    if( (getGameCycleCount() + static_cast<int>(getHouse()->getHouseID())) % AIUPDATEINTERVAL != 0) {
         // we are not updating this AI player this cycle
         return;
     }
